Fix gate level sim
make gpio_defaults writes a modified caravel.v to the
caravel_user_project verilog/gl directory so we need to use that and not
the version in the caravel repo. We also need the gate level
implementations of the various gpio_defaults_block modules (only a few
are created and referenced in the caravel repo).